backend/pyproject.toml
[tool.poetry]
name = "rag-agent-ia"
version = "1.0.0"
description = "Système RAG personnalisé avec Mistral AI, Supabase et sources de données externes"
authors = ["Jimmy Gay <jimmy@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"

# LLM & Embeddings
mistralai = "^1.2.0"
langchain = "^0.3.0"
langchain-mistralai = "^0.2.0"
langchain-community = "^0.3.0"

# Vector Store & Database
supabase = "^2.10.0"
pgvector = "^0.3.0"

# Data Providers
PyGithub = "^2.4.0"
PyMuPDF = "^1.24.0"

# Web Search Agent
httpx = "^0.27.0"

# API Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = {extras = ["email"], version = "^2.9.0"}
pydantic-settings = "^2.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

# UI Testing
streamlit = "^1.40.0"

# Utilities
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
structlog = "^24.4.0"
tiktoken = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
mypy = "^1.13.0"
ruff = "^0.8.0"
black = "^24.10.0"
pre-commit = "^4.0.0"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
# B008: Depends() as default is valid FastAPI pattern
# B904: raise from is optional in many error handlers
# SIM105: try-except-pass is intentional for optional operations
ignore = ["E501", "B008", "B904", "SIM105"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
